Initial import of CRA + Auth and API facade
Bit of an experiment to see if I can connect the client correctly, no
tests just yet.

General design idea is that API clients would have a standard interface,
that generates RdyEvents, the calendar could then be extended to take on
arbitrary calendaring sources.

Coverage in the next couple of commits.