Finish basic requirements (without testing)
This commit includes pretty much the basic baseline of the calendar.
It's still lacking some notable features (included in the README)

Most notably it's missing test coverage for pretty much everything. As I
was mostly remembering how react worked, and how react + hooks works in
general. I determined TDD was a bit combersome for the exploration I was
endeavoring to do.